refactor: rename create-application-harness-options
