+++ packages/spectacular/src/lib/application-testing/application-harness/create-application-harness.ts
@@ -5,16 +5,13 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { SpectacularAppComponent, spectacularAppTag } from '../app-component/spectacular-app.component';
 import { SpectacularAppModule } from '../app-component/spectacular-app.module';
 import { bootstrapComponent } from '../util-bootstrapping/bootstrap-component';
-import { BootstrapSpectacularApplicationOptions } from './bootstrap-spectacular-application-options';
+import { BootstrapSpectacularApplicationOptions } from './create-application-harness-options';
 import { SpectacularApplicationHarness } from './spectacular-application-harness';
 
 /**
  * Bootstrap a test application with the specified metadata. Useful to test
  * configuration Angular modules, bootstrap listeners, and application
  * initializers.
- *
- * @param testModuleMetadata Metadata registered with the Angular testing
- *   module.
  */
 export function createApplicationHarness({
   imports = [],
